fix: recognize the junk create when [terminal] new_cwd is a fixed path
The intercept's junk marker was cwd == the .mirror-pane placeholder, which
holds only under cwd-follow. With a fixed new_cwd the native create lands at
that configured path and the intercept never fired. fixed_new_cwd() reads the
policy from herdr's config and is_bare_placeholder accepts either marker;
herdr_config_path() moves verbatim from binding.rs to util.rs to be shared.
